+++ test/plug/plug_router_test.exs
@@ -44,7 +44,7 @@ defmodule Cldr.Plug.Router.Test do
              language_variant: nil
            }
 
-    assert Cldr.get_current_locale(TestBackend.Cldr) == %Cldr.LanguageTag{
+    assert Cldr.get_locale(TestBackend.Cldr) == %Cldr.LanguageTag{
              canonical_locale_name: "fr-Latn-FR",
              cldr_locale_name: "fr",
              extensions: %{},